# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API PixelCoffee",
    "description": "Api feita para o projeto PixelCoffee",
    "version": "1.3.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/",
      "description": "URL de desenvolvimento"
    }
  ], 
...
  "paths": {
...
    "/menu/takeout-order" : {
      "post": {
        "summary": "Criar pedido para viajem",
        "tags": ["menu"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                  "items": {
                    "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "integer"
                        },
                        "id": {
                          "type": "integer"
                        },
                        "price": {
                          "type": "number"
                        }
                      },
                    "required": ["name", "quantity", "id", "price"]
                }
              }
            }
          }
        }
      }
    },
    "/menu/table-order" : {
      "post": {
        "summary": "Criar pedido para viajem",
        "tags": ["menu"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "table": {
                  "type": "string"
                },
                "type": "array",
                  "items": {
                    "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "integer"
                        },
                        "id": {
                          "type": "integer"
                        },
                        "price": {
                          "type": "number"
                        }
                      },
                    "required": ["name", "quantity", "id", "price"]
                }
              }
            }
          }
        }
      }
    },
...
  }
}